:Release Notes: Fix minor bugs and code refactoring
:Detailed Notes:
- Convert function that generates date format file name to a library
- Modify repeated code into a common function
- Fix the part where the config created by the command is different from the git file
- Sort order of source require

:Testing Performed:
1. Pass unit test on ose
2. Check below commands and results
2-1. ares-config -p ose
    > check git status and no diff files/config/config.json
2-2. Check each operation of ares-install/launch/device

:Issues Addressed:
[WRN-1725] Fix minor bugs and code refactoring of CLI
