+++ bin/ares-device.js
@@ -6,10 +6,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const path = require('path'),
-    async = require('async'),
-    log = require('npmlog'),
+const async = require('async'),
     nopt = require('nopt'),
+    log = require('npmlog'),
+    path = require('path'),
     deviceLib = require('./../lib/device'),
     commonTools = require('./../lib/base/common-tools');
