+++ bin/ares-push.js
@@ -6,10 +6,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const path = require('path'),
+const nopt = require('nopt'),
     log = require('npmlog'),
-    nopt = require('nopt'),
-    Pusher = require('./../lib/pusher'),
+    path = require('path'),
+    Pusher = require('./../lib/pusher');
     commonTools = require('./../lib/base/common-tools'),
     spinner = require('./../lib/util/spinner');
