+++ bin/ares-setup-device.js
@@ -8,12 +8,11 @@
 
 const async = require('async'),
     inquirer = require('inquirer'),
-    log = require('npmlog'),
     nopt = require('nopt'),
+    log = require('npmlog'),
     path = require('path'),
     Ssdp = require('ssdp-js');
-
-const commonTools = require('./../lib/base/common-tools'),
+    commonTools = require('./../lib/base/common-tools'),
     novacom = require('./../lib/base/novacom');
 
 const version = commonTools.version,
@@ -102,6 +101,7 @@ if (argv.list) {
 } else if (argv.search || argv.timeout) {
     op = search;
 } else if (argv.add || argv.modify || argv.info) {
+    // console.log("?")
     op = modifyDeviceInfo;
 } else if (argv.default) {
     op = setDefaultDeviceInfo;
@@ -449,9 +449,7 @@ function _queryDeviceInfo(selDevice, next) {
             if (err) {
                 return next(err);
             }
-            next(null, {
-                "msg": "Success to " + mode + " a device!!"
-            });
+            next();
         });
     });
 }
@@ -623,7 +621,7 @@ function modifyDeviceInfo(next) {
             if (err) {
                 return next(err);
             }
-            next(null, {"msg": "Success to " + mode + " a device named " + inDevice.name + "!!"});
+            next();
         });
     } catch (err) {
         next(err);
@@ -642,7 +640,7 @@ function setDefaultDeviceInfo(next) {
             if (err) {
                 return next(err);
             }
-            next(null, {"msg": "Success to set device named " + argv.default + " to default!!"});
+            next();
         });
     } catch (err) {
         next(err);
@@ -665,7 +663,7 @@ function removeDeviceInfo(next) {
             if (err) {
                 return next(err);
             }
-            next(null, {"msg": "Success to remove a device named " + argv.remove + "!!"});
+            next();
         });
     } catch (err) {
         next(err);
