+++ lib/base/cli-appdata.js
@@ -5,13 +5,13 @@
  */
 
 const fs = require('fs'),
-    log = require('npmlog'),
+    Validator = require('jsonschema').Validator,
     mkdirp = require('mkdirp'),
+    log = require('npmlog'),
     path = require('path'),
     shelljs = require('shelljs'),
-    Validator = require('jsonschema').Validator,
-    clone = require('./../util/objclone').deepCopy,
-    errHndl = require('./error-handler.js');
+    errHndl = require('./error-handler.js')
+    clone = require('./../util/objclone').deepCopy;
 
 const Cli = (function() {
     let instance;
@@ -196,7 +196,7 @@ const Cli = (function() {
         setConfig: function (configData, next) {
             try {
                 this.config = configData;
-                fs.writeFileSync(this.configFile, JSON.stringify(this.config, null, Number(4)));
+                fs.writeFileSync(this.configFile, JSON.stringify(this.config, null, Number(4)) + "\n");
             } catch (err) {
                 return finish(err, this.config, next);
             }
