+++ lib/base/common-tools.js
@@ -4,14 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const help = require('./help-format'),
-    version = require('./version-tools'),
-    errMsg = require('./error-handler'),
+const Appdata = require('./cli-appdata'),
     cliControl = require('./cli-control'),
+    errMsg = require('./error-handler'),
+    help = require('./help-format'),
+    sdkenv = require('./sdkenv'),
     setupDevice = require('./setup-device'),
-    Appdata = require('./cli-appdata'),
-    sdkenv = require('./sdkenv');
-
+    version = require('./version-tools');
 
 (function() {
     const commonTools = {};
