+++ lib/base/help-format.js
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const sprintf = require('sprintf-js').sprintf,
-    fs = require('fs'),
-    path = require('path');
+const fs = require('fs'),
+    path = require('path'),
+    sprintf = require('sprintf-js').sprintf;
 
 (function () {
     const helpFormat = {};
