+++ lib/base/luna.js
@@ -5,8 +5,8 @@
  */
 
 const npmlog = require('npmlog'),
-    Eof = require('../util/eof'),
-    errHndl = require('./error-handler');
+    errHndl = require('./error-handler'),
+    Eof = require('../util/eof');
 
 (function() {
     const log = npmlog;
