+++ lib/base/novacom.js
@@ -4,20 +4,20 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const fs = require('fs'),
-    path = require('path'),
-    util = require('util'),
-    stream = require('stream'),
-    net = require('net'),
-    mkdirp = require('mkdirp'),
-    async = require('async'),
-    log = require('npmlog'),
-    Ssh2 = require('ssh2'),
+const async = require('async'),
     chalk = require('chalk'),
+    fs = require('fs'),
+    mkdirp = require('mkdirp'),
+    net = require('net'),
+    log = require('npmlog'),
+    path = require('path'),
     shelljs = require('shelljs'),
-    server = require('./server'),
+    Ssh2 = require('ssh2'),
+    stream = require('stream'),
+    util = require('util'),
+    Appdata = require('./cli-appdata'),
     errHndl = require('./error-handler'),
-    Appdata = require('./cli-appdata');
+    server = require('./server');
 
 // novacom emulation layer, on top of ssh
 (function() {
