+++ lib/base/server.js
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const fs = require('fs'),
-    Express = require('express'),
-    http = require('http'),
-    bodyParser = require('body-parser'),
+const bodyParser = require('body-parser'),
     spawn = require('child_process').spawn,
+    Express = require('express'),
+    fs = require('fs'),
+    http = require('http'),
     errHndl = require('./error-handler');
 
 (function () {
