+++ lib/base/setup-device.js
@@ -5,8 +5,8 @@
  */
 
 const async = require('async'),
-    Table = require('easy-table'),
     chalk = require('chalk'),
+    Table = require('easy-table'),
     novacom = require('./novacom');
 
 (function () {
