+++ lib/base/version-tools.js
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const semver = require('semver'),
+const fs = require('fs'),
     path = require('path'),
-    fs = require('fs');
+    semver = require('semver');
 
 (function () {
     const vtools = {};
