+++ lib/device.js
@@ -4,21 +4,21 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const util = require('util'),
-    async = require('async'),
-    npmlog = require('npmlog'),
-    fs = require('fs'),
-    path = require('path'),
-    mkdirp = require('mkdirp'),
+const async = require('async'),
     chalk = require('chalk'),
+    fs = require('fs'),
+    mkdirp = require('mkdirp'),
+    npmlog = require('npmlog'),
+    path = require('path'),
+    util = require('util'),
+    pullLib = require('./pull'),
+    errHndl = require('./base/error-handler'),
     luna = require('./base/luna'),
     novacom = require('./base/novacom'),
-    errHndl = require('./base/error-handler'),
-    pullLib = require('./pull'),
-    convertJsonToList = require('./../lib/util/json').convertJsonToList;
+    createDateFileName = require('./util/createFileName').createDateFileName,
+    convertJsonToList = require('./util/json').convertJsonToList;
 
 (function() {
-
     const log = npmlog;
     log.heading = 'device';
     log.level = 'warn';
@@ -55,13 +55,7 @@ const util = require('util'),
             });
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _getOsInfo(next) {
@@ -209,13 +203,7 @@ const util = require('util'),
             });
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _getSessionList(next) {
@@ -269,13 +257,7 @@ const util = require('util'),
             });
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _makeCaptureOption(next) {
@@ -285,7 +267,7 @@ const util = require('util'),
                     captureFormat = "PNG"; // PNG is default format
 
                 if (options.outputPath === null){
-                    captureFileName = _createDateFileName();
+                    captureFileName = options.session.target.name + "_" + "display" + options.display + "_" + createDateFileName("") + ".png";
                     destinationPath = path.resolve('.');
                 } else {
                     // Get directory path, file path, file Extenstion from given path
@@ -309,7 +291,7 @@ const util = require('util'),
                         destinationPath = path.resolve(parseDirPath);
                     } else if (parseBase && (parseExt === "")) {
                         // it's directory path. Use it as destination dir Path
-                        captureFileName = _createDateFileName();
+                        captureFileName = options.session.target.name + "_" + "display" + options.display + "_" + createDateFileName("") + ".png";
                         destinationPath = path.resolve(options.outputPath);
                     }
                 }
@@ -323,31 +305,6 @@ const util = require('util'),
                 options.silent = true;
                 
                 next();
-            }
-
-            function _createDateFileName() {
-                // Capture file name generate : targetname_displayX_yyyymmddhhmmss.png"
-                function __getDate(){
-                    const curDate = new Date();
-                    const dateFormat = __pad(curDate.getFullYear(), 2)
-                        +  __pad((curDate.getMonth()+1), 2)
-                        + __pad(curDate.getDate(), 2)
-                        + __pad(curDate.getHours(), 2)
-                        + __pad(curDate.getMinutes(), 2)
-                        + __pad(curDate.getSeconds(), 2);
-                    return dateFormat;
-                }
-
-                function __pad(number, length) {
-                    let str = '' + number;
-                    while (str.length < length) {
-                        str = '0' + str;
-                    }
-                    return str;
-                }
-
-                const fileName = options.session.target.name + "_" + "display" + options.display + "_" + __getDate() + ".png";
-                return fileName;
             }
 
             function _captureScreenFile(next) {
@@ -407,6 +364,17 @@ const util = require('util'),
             }
         }
     };
+
+    function makeSession(options, next){
+        log.info("device#makeSession");
+        options.nReplies = 1; // -n 1
+        if (!options.session) {
+            const printTarget = true;
+            options.session = new novacom.Session(options.device, printTarget, next);
+        } else {
+            next();
+        }
+    }
 
     if (typeof module !== 'undefined' && module.exports) {
         module.exports = device;
