+++ lib/generator.js
@@ -6,14 +6,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const path = require('path'),
-    promise = require('bluebird'),
-    log = require('npmlog'),
+const promise = require('bluebird'),
     Table = require('easy-table'),
-    errHndl = require('./base/error-handler'),
     fs = promise.promisifyAll(require('fs-extra')),
-    readJsonSync = require('./util/json').readJsonSync,
+    log = require('npmlog'),
+    path = require('path'),
+    errHndl = require('./base/error-handler'),
     copyToDirAsync = require('./util/copy').copyToDirAsync,
+    readJsonSync = require('./util/json').readJsonSync,
     merge = require('./util/merge');
 
 function Generator(tmplFile) {
