+++ lib/inspect.js
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const util = require('util'),
-    async = require('async'),
-    path = require('path'),
+const async = require('async'),
     npmlog = require('npmlog'),
+    path = require('path'),
     request = require('request'),
+    util = require('util'),
     installer = require('./install'),
     launcher = require('./launch'),
     errHndl = require('./base/error-handler'),
