+++ lib/install.js
@@ -4,13 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const fs = require('fs'),
-    path = require('path'),
-    util = require('util'),
-    npmlog = require('npmlog'),
-    async = require('async'),
-    streamBuffers = require('stream-buffers'),
+const async = require('async'),
     crypto = require('crypto'),
+    fs = require('fs'),
+    npmlog = require('npmlog'),
+    path = require('path'),
+    streamBuffers = require('stream-buffers'),
+    util = require('util'),
     sessionLib = require('./session'),
     Appdata = require('./base/cli-appdata'),
     errHndl = require('./base/error-handler'),
@@ -74,12 +74,7 @@ const fs = require('fs'),
             async.waterfall([
                 function(next) {
                     options.nReplies = 0; // -i
-                    if (!options.session) {
-                        const printTarget = true;
-                        options.session = new novacom.Session(options.device, printTarget, next);
-                    } else {
-                        next(null, options.session);
-                    }
+                    makeSession(options, next);
                 },
                 function(session, next) {
                     if (options.opkg) {
@@ -271,13 +266,8 @@ const fs = require('fs'),
             options = options || {};
             async.waterfall([
                 function(next) {
-                    options.nReplies = undefined; // -i
-                    if (!options.session) {
-                        const printTarget = true;
-                        options.session = new novacom.Session(options.device, printTarget, next);
-                    } else {
-                        next(null, options.session);
-                    }
+                    options.nReplies = 0; // -i
+                    makeSession(options, next);
                 },
                 function(session, next) {
                     if (options.opkg) {
@@ -381,12 +371,7 @@ const fs = require('fs'),
             async.series([
                 function(next) {
                     options.nReplies = 1; // -n 1
-                    if (!options.session) {
-                        const printTarget = true;
-                        options.session = new novacom.Session(options.device, printTarget, next);
-                    } else {
-                        next();
-                    }
+                    makeSession(options, next);
                 },
                 function(next) {
                     if (options.opkg) {
@@ -463,6 +448,16 @@ const fs = require('fs'),
             });
         }
     };
+
+    function makeSession(options, next){
+        log.info("install#makeSession");
+        if (!options.session) {
+            const printTarget = true;
+            options.session = new novacom.Session(options.device, printTarget, next);
+        } else {
+            next(null, options.session);
+        }
+    }
 
     if (typeof module !== 'undefined' && module.exports) {
         module.exports = installer;
