+++ lib/launch.js
@@ -4,17 +4,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const util = require('util'),
-    async = require('async'),
+const async = require('async'),
     npmlog = require('npmlog'),
-    luna = require('./base/luna'),
+    os = require( 'os' ),
     path = require('path'),
+    util = require('util'),
     inspector = require('./inspect'),
     installer = require('./install'),
-    novacom = require('./base/novacom'),
-    errHndl = require('./base/error-handler'),
     sessionLib = require('./session'),
-    os = require( 'os' );
+    errHndl = require('./base/error-handler'),
+    luna = require('./base/luna'),
+    novacom = require('./base/novacom');
 
 (function() {
     const log = npmlog;
@@ -108,13 +108,7 @@ const util = require('util'),
             }
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _runAppServer(next){
@@ -245,13 +239,7 @@ const util = require('util'),
             });
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _getSessionList(next) {
@@ -321,13 +309,7 @@ const util = require('util'),
             });
 
             function _makeSession(next) {
-                options.nReplies = 1; // -n 1
-                if (!options.session) {
-                    const printTarget = true;
-                    options.session = new novacom.Session(options.device, printTarget, next);
-                } else {
-                    next();
-                }
+                makeSession(options, next);
             }
 
             function _getSessionList(next) {
@@ -368,6 +350,17 @@ const util = require('util'),
             }
         }
     };
+
+    function makeSession(options, next){
+        log.info("launch#makeSession");
+        options.nReplies = 1; // -n 1
+        if (!options.session) {
+            const printTarget = true;
+                    options.session = new novacom.Session(options.device, printTarget, next);
+        } else {
+            next();
+        }
+    }
 
     function checkDisplayAffinity(options, params, next) {
         // case of do not need to call session call(ose), check displayAffinity with display
