+++ lib/log.js
@@ -11,7 +11,8 @@ const util = require('util'),
     path = require('path'),
     novacom = require('./base/novacom'),
     errHndl = require('./base/error-handler'),
-    sessionLib = require('./session');
+    sessionLib = require('./session'),
+    createDateFileName = require('./util/createFileName').createDateFileName;
 
 (function() {
 
@@ -339,7 +340,7 @@ const util = require('util'),
     function createLogFile(options, next) {
         log.info("log#createLogFile");
         if (options.argv.argv.remain.length === 0) {
-            options.argv.argv.remain[0] = path.resolve(_createDefaultFileName());
+            options.argv.argv.remain[0] = path.resolve(createDateFileName("_")+".log");
         } else {
             if (path.extname(options.argv.argv.remain[0]) === "") {
                 options.argv.argv.remain[0] = options.argv.argv.remain[0] + ".log";
@@ -357,32 +358,6 @@ const util = require('util'),
 
         log.info("options.argv.argv.remain[0]: " + options.argv.argv.remain[0]);
         next();
-    }
-
-    function _createDefaultFileName() {
-        log.info("log#_createDefaultFileName");
-        // Default log file name generate : yyyymmdd_hhmmss.log"
-        return __getDate() + ".log";
-
-        function __getDate(){
-            const curDate = new Date();
-            const dateFormat = __pad(curDate.getFullYear(), 2)
-                +  __pad((curDate.getMonth()+1), 2)
-                + __pad(curDate.getDate(), 2)
-                + "_"
-                + __pad(curDate.getHours(), 2)
-                + __pad(curDate.getMinutes(), 2)
-                + __pad(curDate.getSeconds(), 2);
-            return dateFormat;
-        }
-
-        function __pad(number, length) {
-            let str = '' + number;
-            while (str.length < length) {
-                str = '0' + str;
-            }
-            return str;
-        }
     }
 
     if (typeof module !== 'undefined' && module.exports) {
