+++ lib/package.js
@@ -5,14 +5,16 @@
  */
 
 const async = require('async'),
-    CombinedStream = require('combined-stream'),
     chardet = require('chardet'),
+    CombinedStream = require('combined-stream'),
+    crypto = require('crypto'),
     ElfParser = require('elfy').Parser,
     encoding = require('encoding'),
     fs = require('fs'),
     fstream = require('fstream'),
-    log = require('npmlog'),
+    Validator = require('jsonschema').Validator,
     mkdirp = require('mkdirp'),
+    log = require('npmlog'),
     path = require('path'),
     rimraf = require('rimraf'),
     shelljs = require('shelljs'),
@@ -20,9 +22,7 @@ const async = require('async'),
     temp = require('temp'),
     uglify = require('terser'),
     util = require('util'),
-    Validator = require('jsonschema').Validator,
     zlib = require('zlib'),
-    crypto = require('crypto'),
     tarFilterPack = require('./tar-filter-pack'),
     errHndl = require('./base/error-handler');
