+++ lib/pull.js
@@ -4,17 +4,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const fs = require('fs'),
-    async = require('async'),
+const async = require('async'),
+    fs = require('fs'),
+    mkdirp = require('mkdirp'),
     npmlog = require('npmlog'),
-    util = require('util'),
     path = require('path'),
     streamBuffers = require('stream-buffers'),
-    mkdirp = require('mkdirp'),
+    util = require('util'),
     errHndl = require('./base/error-handler'),
     novacom = require('./base/novacom'),
     spinner = require('./util/spinner');
-
+    
 (function() {
     const log = npmlog;
     log.heading = 'pull';
