+++ lib/pusher.js
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const fs = require('fs'),
-    async = require('async'),
+const async = require('async'),
+    fs = require('fs'),
     log = require('npmlog'),
-    util = require('util'),
     path = require('path'),
     streamBuffers = require('stream-buffers'),
+    util = require('util'),
     errHndl = require('./base/error-handler'),
     novacom = require('./base/novacom'),
     spinner = require('./util/spinner');
