+++ lib/session.js
@@ -5,8 +5,8 @@
  */
 
 const npmlog = require('npmlog'),
-    luna = require('./base/luna'),
-    errHndl = require('./base/error-handler');
+errHndl = require('./base/error-handler'),
+    luna = require('./base/luna');
 
 (function() {
     const log = npmlog;
