+++ lib/shell.js
@@ -6,9 +6,9 @@
 
 const async = require('async'),
     npmlog = require('npmlog'),
-    novacom = require('./base/novacom'),
+    sessionLib = require('./session'),
     errHndl = require('./base/error-handler'),
-    sessionLib = require('./session');
+    novacom = require('./base/novacom');
 
 (function() {
     const log = npmlog;
@@ -23,13 +23,7 @@ const async = require('async'),
             log.info('shell#remoteRun()');
             async.series([
                 function(next) {
-                    options.nReplies = 1;
-                    if (!options.session) {
-                        const printTarget = false;
-                        options.session = new novacom.Session(options.device, printTarget, next);
-                    } else {
-                        next();
-                    }
+                    makeSession(options, next);
                 },
                 function(next) {
                     if (options && options.display) {
@@ -70,12 +64,7 @@ const async = require('async'),
                 log.info('shell#shell()');
                 async.series([
                     function(next) {
-                        if (!session) {
-                            const printTarget = false;
-                            session = new novacom.Session(options.device, printTarget, next);
-                        } else {
-                            setImmediate(next);
-                        }
+                        makeSession(options, next);
                     },
                     function(next) {
                         const winOpts = {
@@ -147,13 +136,7 @@ const async = require('async'),
 
             async.series([
                 function(next) {
-                    options.nReplies = 1;
-                    if (!options.session) {
-                        const printTarget = false;
-                        options.session = new novacom.Session(options.device, printTarget, next);
-                    } else {
-                        next();
-                    }
+                    makeSession(options, next);
                 },
                 function(next) {
                     if (options && options.display) {
@@ -176,6 +159,17 @@ const async = require('async'),
             });
         }
     };
+
+    function makeSession(options, next){
+        log.info("shell#makeSession");
+        options.nReplies = 1; // -n 1
+        if (!options.session) {
+            const printTarget = false;
+            options.session = new novacom.Session(options.device, printTarget, next);
+        } else {
+            next();
+        }
+    }
 
     if (typeof module !== 'undefined' && module.exports) {
         module.exports = shell;
