+++ lib/tar-filter-pack.js
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-const path = require('path'),
+const collect = require('fstream').collect,
     inherits = require('inherits'),
-    tar = require('tar'),
-    collect = require('fstream').collect;
+    path = require('path'),
+    tar = require('tar');
 
 module.exports = TarFilterPack;
 inherits(TarFilterPack, tar.Pack);
